Create separate component for TimerRender that is passed to main timer component
