+++ src/views/GuildPad/Pad/Cards/index.tsx
@@ -9,18 +9,19 @@ import { Socials, GuildpadConfig, GUILDPAD_STATUS, TYPE } from 'config/constants
 import { Globe, Send, Twitter } from 'react-feather'
 import { SiDiscord, SiYoutube } from 'react-icons/si'
 import { useFetchBanner, useFetchPadBG } from 'utils/assetFetch'
-import Timer from 'views/GuildPad/components/Timer'
+import TimerRenderer from 'views/GuildPad/components/TimerRenderer'
+import Timer from 'components/Launchpad/Timer'
+import Content from 'components/Launchpad/Content'
 import { getStatus } from 'utils/guildpadHelpers'
 import Anchor from 'components/Launchpad/Anchor'
 import SvgIcon from 'components/Launchpad/SvgIcon'
 import { ReactComponent as MediumIcon } from 'components/Launchpad/icons/MediumIcon.svg'
 import { Card as SCard, CardHeader as SCardHeader, Text, Heading, Flex, Button } from '@metagg/mgg-uikit'
 import TokenLogo from 'components/Launchpad/Logo'
-import { PostBody, NavOption, SaleContainer, SaleRow } from '../../../../components/Launchpad/styled'
+// import { PostBody, NavOption, SaleContainer, SaleRow } from '../../../../components/Launchpad/styled'
 import Boxcard from '../BoxCard'
-import '../../../../css/styleFX.css'
-import Content from '../../../../components/Launchpad/Content'
 import IdoCard from './Ido'
+
 
 const GCard = styled(SCard)<{ src?: string }>`
   border: 2px solid ${({ theme }) => theme.colors.MGG_active};
@@ -130,63 +131,29 @@ const ContainerProjDesc = styled(Flex)`
   position: relative;
   margin: 1.5rem;
 `
-const Box = styled.div`
-  height: 100%;
-  min-width: 100px;
-`
+// const Box = styled.div`
+//   height: 100%;
+//   min-width: 100px;
+// `
 
-const HOrbitron = styled(Heading)`
-  display: flex;
-  justify-content: center;
-  text-align: center;
-  font-family: Orbitron !important;
-  line-height: 0.7em;
-  color: grey;
-`
-const TimeSpan = styled.div`
-  margin: 1rem 0 1.5rem 0;
-`
+// const HOrbitron = styled(Heading)`
+//   display: flex;
+//   justify-content: center;
+//   text-align: center;
+//   font-family: Orbitron !important;
+//   line-height: 0.7em;
+//   color: grey;
+// `
+// const TimeSpan = styled.div`
+//   margin: 1rem 0 1.5rem 0;
+// `
 
 // COUNTDOWN TIMER
 const CountDown: React.FC<{ round: string, start?:boolean, end?:number}> = ({round, start, end}) => {
   const endDate = end
   const isStart = start;
 
-  const Renderer = (days?: number, hours?: number, minutes?: number, seconds?: number) => {
-    return(
-      <div>
-      <Heading color='white' style={{textAlign: 'center', paddingTop: '0.5rem'}} size="l">ROUND {round} ENDS IN</Heading>
-      <TimerContainer>
-      <TimerBox>
-          <Box>
-            <TimeSpan>
-              <HOrbitron size="xl" className='glow'>{days}</HOrbitron>
-            </TimeSpan>
-            <Text color='white' fontSize="1rem"> DAYS </Text>
-          </Box>
-          <Box>
-            <TimeSpan>
-              <HOrbitron size="xl" className='glow'>{hours}</HOrbitron>
-            </TimeSpan>
-            <Text color='white' fontSize="1rem"> HOURS </Text>
-          </Box>
-          <Box>
-            <TimeSpan>
-              <HOrbitron size="xl" className='glow'>{minutes}</HOrbitron>
-            </TimeSpan>
-            <Text color='white' fontSize="1rem"> MINUTES</Text>
-          </Box>
-          <Box>
-            <TimeSpan>
-              <HOrbitron size="xl" className='glow'>{seconds}</HOrbitron>
-            </TimeSpan>
-            <Text color='white' fontSize="1rem"> SECONDS</Text>
-          </Box>
-        </TimerBox>
-      </TimerContainer>
-      </div>
-    )
-   }
+  const Renderer = (days?: number, hours?: number, minutes?: number, seconds?: number) => <TimerRenderer days={days} hours={hours} minutes={minutes} seconds={seconds} round={round} />
 
   return (
     <TimerContainer justifyContent="right" padding={isStart? '10px':'0px'}>
